Fix bug where SG start date would not display
The SG start date should only display if the current date is on or after
that start date. However, the original time comparison logic did not
respect the site's chosen timezone, resulting in the start date hiding
prematurely in some cases (because US times are relatively earlier than
GMT+0 times).