Fix bug where header nav would duplicate
In particular, the header navigation menu would duplicate itself in the
Customizer preview when edited. After some investigation, this bug was
found to be caused by the fact that the navigation container was created
outside of the wp_nav_menu() function, thereby breaking some assumption
the Customizer was making when performing the partial refresh.